+++ plugin/src/net/groboclown/idea/p4ic/changes/P4ChangeListIdImpl.java
@@ -136,6 +136,6 @@ class P4ChangeListIdImpl implements P4ChangeListId {
 
     @Override
     public String toString() {
-        return getServerConfigId() + "+" + getClientName() + "@" + getChangeListId();
+        return getServerConfigId() + "/" + getClientName() + "@" + getChangeListId();
     }
 }