+++ CHANGES.md
@@ -16,6 +16,9 @@
       and when the files are considered out of sync.
     * Cached the jobs loaded from the server.
     * Added a monitor to track the number of actual server calls.
+    * Capped the number of simultaneous connections to the same server/workspace (2).
+    * Capped the frequency at which changelists will be refreshed (one per second).
+      This is due to editing files triggering a changelist refresh, which can become quite noisy.
 * Bug fixes.
     * Reduced the amount of log messages.
     * Eliminated NPE related to changelist caching.